+++ src/classes/DBHelper/BaseCollection/Keys/Key.php
@@ -2,9 +2,12 @@
 
 declare(strict_types=1);
 
+use Application\AppFactory;
+use AppUtils\Microtime;
+use DBHelper\BaseRecord\BaseRecordException;
+
 class DBHelper_BaseCollection_Keys_Key
 {
-    public const ERROR_CANNOT_GENERATE_VALUE = 87601;
 
     /**
      * @var DBHelper_BaseCollection_Keys
@@ -89,16 +92,28 @@ class DBHelper_BaseCollection_Keys_Key
      * 2. The full data set being validated (for lookups)
      * 3. The key instance
      *
-     * @param callable $callback
+     * @param callable(mixed, array<string,mixed>, DBHelper_BaseCollection_Keys_Key) : bool $callback
      * @return $this
-     * @throws Application_Exception
      */
-    public function setValidation($callback) : DBHelper_BaseCollection_Keys_Key
+    public function setValidation(callable $callback) : self
     {
-        Application::requireCallableValid($callback);
-
         $this->validation = $callback;
         return $this;
+    }
+
+    /**
+     * Sets a regular expression validation for the key.
+     *
+     * The regex must be a full regex, including delimiters.
+     *
+     * @param string $regex
+     * @return $this
+     */
+    public function setRegexValidation(string $regex) : self
+    {
+        return $this->setValidation(function(mixed $value) use ($regex) : bool {
+            return is_string($value) && preg_match($regex, $value);
+        });
     }
 
     /**
@@ -130,7 +145,7 @@ class DBHelper_BaseCollection_Keys_Key
      *
      * The method must return the generated value.
      *
-     * @param callable $callback
+     * @param callable(DBHelper_BaseCollection_Keys_Key, array<string,mixed>): mixed $callback
      * @return $this
      */
     public function setGenerator(callable $callback) : DBHelper_BaseCollection_Keys_Key
@@ -153,19 +168,19 @@ class DBHelper_BaseCollection_Keys_Key
      * @throws DBHelper_Exception
      *
      * @see DBHelper_BaseCollection_Keys_Key::setGenerator()
-     * @see DBHelper_BaseCollection_Keys_Key::ERROR_CANNOT_GENERATE_VALUE
+     * @see BaseRecordException::ERROR_CANNOT_GENERATE_KEY_VALUE
      */
-    public function generateValue(array $data)
+    public function generateValue(array $data) : mixed
     {
         if(isset($this->generator))
         {
             return call_user_func($this->generator, $this, $data);
         }
 
-        throw new DBHelper_Exception(
+        throw new BaseRecordException(
             'Cannot generate value, no generator present',
             'No callback has been set for generating the value.',
-            self::ERROR_CANNOT_GENERATE_VALUE
+            BaseRecordException::ERROR_CANNOT_GENERATE_KEY_VALUE
         );
     }
 
@@ -175,7 +190,7 @@ class DBHelper_BaseCollection_Keys_Key
      * @param mixed $value
      * @param array<string,mixed> $dataSet The full data set being used, for looking up other values for the validation.
      */
-    public function validate($value, array $dataSet) : void
+    public function validate(mixed $value, array $dataSet) : void
     {
         if($this->validation === null)
         {
@@ -183,5 +198,70 @@ class DBHelper_BaseCollection_Keys_Key
         }
 
         call_user_func($this->validation, $value, $dataSet, $this);
+    }
+
+    public function setMicrotimeGenerator() : self
+    {
+        $this->setMicrotimeValidation();
+
+        return $this->setGenerator(function() : Microtime {
+            return Microtime::createNow();
+        });
+    }
+
+    public function setMicrotimeValidation() : self
+    {
+        return $this->setValidation(function(mixed $value) : bool
+        {
+            if($value instanceof Microtime) {
+                return true;
+            }
+
+            if(is_string($value)) {
+                try {
+                    Microtime::createFromString($value);
+                    return true;
+                } catch(Throwable) {
+                    return false;
+                }
+            }
+
+            return false;
+        });
+    }
+
+    public function setUserValidation() : self
+    {
+        return $this->setValidation(function(mixed $value) : bool
+        {
+            if(is_string($value) && is_numeric($value)) {
+                $value = (int)$value;
+            }
+
+            return is_int($value) && $value > 0 && AppFactory::createUsers()->idExists($value);
+        });
+    }
+
+    /**
+     * @param string[] $allowedValues
+     * @return $this
+     */
+    public function setEnumValidation(array $allowedValues) : self
+    {
+        return $this->setValidation(function(mixed $value) use ($allowedValues) : bool
+        {
+            if(is_numeric($value)) {
+                $value = (string)$value;
+            }
+
+            return in_array($value, $allowedValues, true);
+        });
+    }
+
+    public function setCurrentUserGenerator() : self
+    {
+        return $this->setGenerator(function() : int {
+            return AppFactory::createUser()->getID();
+        });
     }
 }